+++ testing/unit_tests/test_testing.py
@@ -5,11 +5,12 @@ from pathlib import Path
 
 from ivadomed.loader.bids_dataframe import BidsDataframe
 from ivadomed import metrics as imed_metrics
-from ivadomed import transforms as imed_transforms
 from ivadomed import utils as imed_utils
 from ivadomed import testing as imed_testing
 from ivadomed import models as imed_models
 from ivadomed.loader import utils as imed_loader_utils, loader as imed_loader
+from ivadomed.transforms.compose import Compose
+from ivadomed.transforms.undo_compose import UndoCompose
 import logging
 from testing.unit_tests.t_utils import create_tmp_dir, __data_testing_dir__, __tmp_dir__, download_data_testing_test_files, path_repo_root
 from testing.common_testing_util import remove_tmp_dir
@@ -81,7 +82,7 @@ def test_inference(download_data_testing_test_files, transforms_dict, test_lst, 
                              num_workers=0)
 
     # Undo transform
-    val_undo_transform = imed_transforms.UndoCompose(imed_transforms.Compose(transforms_dict))
+    val_undo_transform = UndoCompose(Compose(transforms_dict))
 
     # Update testing_params
     testing_params.update({
@@ -181,7 +182,7 @@ def test_inference_2d_microscopy(download_data_testing_test_files, transforms_di
                              num_workers=0)
 
     # Undo transform
-    val_undo_transform = imed_transforms.UndoCompose(imed_transforms.Compose(transforms_dict))
+    val_undo_transform = UndoCompose(Compose(transforms_dict))
 
     # Update testing_params
     testing_params.update({
@@ -274,7 +275,7 @@ def test_inference_target_suffix(download_data_testing_test_files, transforms_di
                              num_workers=0)
 
     # Undo transform
-    val_undo_transform = imed_transforms.UndoCompose(imed_transforms.Compose(transforms_dict))
+    val_undo_transform = UndoCompose(Compose(transforms_dict))
 
     # Update testing_params
     testing_params.update({
